feat(filters): Implement filters using decorators fix(scheduler): use `node-schedule` instead of `cron` for stability reasons test(scheduler): Add a test for date-based scheduling
Filters are used to modify behavior of a function, there are two types of filters,
`preprocessor`s and `postprocessor`s.

`preprocessors` are used to modify a function's arguments while
`postprocessors` are used to modify a function's return value.

A new `@filterable` decorator has been implemented to allow a function to be
filtered using processors. Filters are ran in order they were registered, we might
consider adding a priority parameter later to give more control over the order
in which the filters are executed.

The `Filters` object's methods are responsible for registering and removing
filters. Filters are attached to functions by a name, the name is chosen by
`@filterable` decorator's first argument.

`postprocess` and `preprocess`, functions used to register filters, return an index
number which can later be used to remove a filter.

The remove method takes the filter name and the index to remove the filter.